Add version 2.0 of programming language posters
These posters were created in Adobe InDesign, and the added
folder contains the export from the 'Package' command. It
includes the source files and all linked files for anyone
to recreate the posters.

We wish to convert this file to an open format at a later
time, to enable greater accessibility.